feat: schemas validation

+++ src/controllers/products.controller.ts
@@ -1,11 +1,16 @@
 import { Request, Response } from "express";
 import { STATUS_CODE } from "../enums/status.code.js";
-import { InsertProduct, ListProducts, ParamsQuery, TotalProducts } from "../protocols/protocols.js";
+import { InsertProduct, ListProducts, QueryCategory, TotalProducts } from "../protocols/protocols.js";
 import * as productsRepository from "../repositories/products.repository.js";
 
 async function createProduct(req: Request, res: Response) {
 	const product = req.body as InsertProduct;
 	try {
+		const nameExists = (await productsRepository.selectProductByName(product.name)).rows[0];
+		if (nameExists) {
+			return res.sendStatus(STATUS_CODE.CONFLICT);
+		}
+
 		await productsRepository.insertProduct(product);
 		return res.sendStatus(STATUS_CODE.CREATED);
 	} catch (error) {
@@ -16,15 +21,17 @@ async function createProduct(req: Request, res: Response) {
 async function productSold(req: Request, res: Response) {
 	const name: string = req.body.name;
 	try {
-		await productsRepository.insertSoldProduct(name);
-
 		const product = (await productsRepository.selectProductByName(name)).rows[0];
+		if (!product) {
+			return res.sendStatus(STATUS_CODE.NOT_FOUND);
+		}
 		if (product.quantity > 0) {
 			await productsRepository.updateProductQuantity(name);
 		} else {
 			return res.status(STATUS_CODE.BAD_REQUEST).send({ message: "Produto esgotado" });
 		}
 
+		await productsRepository.insertSoldProduct(name);
 		return res.sendStatus(STATUS_CODE.OK);
 	} catch (error) {
 		return res.sendStatus(STATUS_CODE.SERVER_ERROR);
@@ -32,18 +39,15 @@ async function productSold(req: Request, res: Response) {
 }
 
 async function getProducts(req: Request, res: Response) {
+	const query = req.query as QueryCategory;
 	try {
-		const products = (await productsRepository.listProducts()).rows as ListProducts[];
-		return res.status(STATUS_CODE.OK).send(products);
-	} catch (error) {
-		return res.sendStatus(STATUS_CODE.SERVER_ERROR);
-	}
-}
+		const category: string = query.category;
+		if (category) {
+			const products = (await productsRepository.listProductsByCategory(category)).rows as ListProducts[];
+			return res.status(STATUS_CODE.OK).send(products);
+		}
 
-async function getProductsByCategory(req: Request, res: Response) {
-	const category: string = req.params.category;
-	try {
-		const products = (await productsRepository.listProductsByCategory(category)).rows as ListProducts[];
+		const products = (await productsRepository.listProducts()).rows as ListProducts[];
 		return res.status(STATUS_CODE.OK).send(products);
 	} catch (error) {
 		return res.sendStatus(STATUS_CODE.SERVER_ERROR);
@@ -54,26 +58,18 @@ async function getProductById(req: Request, res: Response) {
 	const id: number = Number(req.params.id);
 	try {
 		const product = (await productsRepository.selectProductById(id)).rows[0] as ListProducts;
+		if (!product) {
+			return res.sendStatus(STATUS_CODE.NOT_FOUND);
+		}
+
 		return res.status(STATUS_CODE.OK).send(product);
 	} catch (error) {
 		return res.sendStatus(STATUS_CODE.SERVER_ERROR);
 	}
 }
 
-async function totalProducts(req: Request, res: Response) {
-	try {
-		const totalAvailable = (await productsRepository.totalAllProductsAvailable()).rows[0] as TotalProducts;
-		const totalSold = (await productsRepository.totalAllProductsSold()).rows[0] as TotalProducts;
-
-		const total: number = Number(totalAvailable.total) + Number(totalSold.total);
-		return res.status(STATUS_CODE.OK).send({ total });
-	} catch (error) {
-		return res.sendStatus(STATUS_CODE.SERVER_ERROR);
-	}
-}
-
 async function totalProductsAvailable(req: Request, res: Response) {
-	const query = req.query as ParamsQuery;
+	const query = req.query as QueryCategory;
 	try {
 		const category: string = query.category;
 		if (category) {
@@ -89,17 +85,11 @@ async function totalProductsAvailable(req: Request, res: Response) {
 }
 
 async function totalProductsSold(req: Request, res: Response) {
-	const query = req.query as ParamsQuery;
+	const query = req.query as QueryCategory;
 	try {
-		const product: string = query.product;
-		if (product) {
-			const total = (await productsRepository.totalProductsSoldByName()).rows as TotalProducts[];
-			return res.status(STATUS_CODE.OK).send(total);
-		}
-
 		const category: string = query.category;
 		if (category) {
-			const total = (await productsRepository.totalProductsSoldByCategory()).rows as TotalProducts[];
+			const total = (await productsRepository.totalProductsSoldByCategory(category)).rows[0] as TotalProducts;
 			return res.status(STATUS_CODE.OK).send(total);
 		}
 
@@ -114,6 +104,11 @@ async function updateProduct(req: Request, res: Response) {
 	const quantity: number = req.body.quantity;
 	const id: number = Number(req.params.id);
 	try {
+		const product = (await productsRepository.selectProductById(id)).rows[0] as ListProducts;
+		if (!product) {
+			return res.sendStatus(STATUS_CODE.NOT_FOUND);
+		}
+
 		await productsRepository.updateProduct(quantity, id);
 		return res.sendStatus(STATUS_CODE.OK);
 	} catch (error) {
@@ -124,6 +119,11 @@ async function updateProduct(req: Request, res: Response) {
 async function deleteProduct(req: Request, res: Response) {
 	const id: number = Number(req.params.id);
 	try {
+		const product = (await productsRepository.selectProductById(id)).rows[0] as ListProducts;
+		if (!product) {
+			return res.sendStatus(STATUS_CODE.NOT_FOUND);
+		}
+
 		await productsRepository.deleteProductById(id);
 		return res.sendStatus(STATUS_CODE.OK);
 	} catch (error) {
@@ -135,9 +135,7 @@ export {
 	createProduct,
 	productSold,
 	getProducts,
-	getProductsByCategory,
 	getProductById,
-	totalProducts,
 	totalProductsAvailable,
 	totalProductsSold,
 	updateProduct,
